pyproject.toml
[project]
name = "whisper-api"
version = "1.0.0"
description = "FastAPI service for OpenAI Whisper speech transcription"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "openai-whisper>=20231117",
    "torch",
    "torchaudio", 
    "ffmpeg-python>=0.2.0",
    "aiofiles>=23.2.0",
    "python-dotenv>=1.0.0",
    "celery>=5.3.4",
    "redis>=5.0.1",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1", 
    "black>=23.11.0",
    "flake8>=6.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
